requirements.txt
requests>=2.31.0
dnspython>=2.4.2
colorama>=0.4.6
beautifulsoup4>=4.12.2
python-whois>=0.8.0
python-nmap>=0.7.1
waybackpy>=3.0.6
psutil>=5.9.0
